fea: Video 컴포넌트 내 query 파일, custom hook 으로 분리
